Moved modern version into count-input-lines2.c and craeted count-input-lines.c which uses the book's syntax.

+++ The-ANSI-C-Programming-Language/Chapter01/count-input-lines2.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <inttypes.h>
 
 int main(void)
 {
-	long currentChar, lineCount;
+	int32_t currentChar, lineCount;
 	lineCount = 0;
 
 	printf("This program will count lines using the newline character.\n");
@@ -13,6 +14,6 @@ int main(void)
 			lineCount++;
 		}
 	}
-	printf("The number of lines types was %ld\n", lineCount);
+	printf("The number of lines types was %d\n", lineCount);
 	return 0;
 }
